feat: Add profile view and user service
Adds a new profile view and user service to the application. The profile view allows users to view and edit their profile information. The user service provides methods for fetching and updating user data. This commit also renames the `AuthException` class to `IdentityException` in the `models/exceptions` directory. The `IdentityException` class is used to represent errors related to identity/authentication processes in the application.